Fix some small todo's

+++ modules/base/src/main/java/io/github/fabricators_of_create/porting_lib/enchant/EnchantmentBonusBlock.java
@@ -12,7 +12,7 @@ public interface EnchantmentBonusBlock {
 	 * @param pos Block position in level
 	 * @return The amount of enchanting power this block produces.
 	 */
-	default float getEnchantPowerBonus(BlockState state, LevelReader level, BlockPos pos) {
+	default int getEnchantPowerBonus(BlockState state, LevelReader level, BlockPos pos) {
 		return state.is(BlockTags.ENCHANTMENT_POWER_PROVIDER) ? 1 : 0;
 	}
 }
